requirements.txt
numpy
pandas
python-dotenv
mysql-connector-python
pymysql
scikit-learn
cryptography
seaborn
catboost
xgboost
#-e .
